(1)修复DefferedShading下点光源的边缘截断问题  (2)后处理增加ACES的ToneMap

+++ SDEngine/Source/GerstnerWave.cpp
@@ -40,7 +40,7 @@ void GerstnerWave::CalculateVertexPos(float time)
 			//pos
 			mWaveVertexData[index].pos.x = fGerstnerX;
 			mWaveVertexData[index].pos.z = fGerstnerZ;
-			mWaveVertexData[index].pos.y = (float)GetWaveVertexHeight(fPositionX, fPositionZ, time);
+			mWaveVertexData[index].pos.y = (float)GetWaveVertexHeight((float)fPositionX, (float)fPositionZ, time);
 			++index;
 		}
 	}
